Separated Contract type tests into its own test AT-7391
